moved delete validator
