+++ x/stakeibc/keeper/msg_server.go
@@ -282,6 +282,19 @@ func (k msgServer) AddValidators(goCtx context.Context, msg *types.MsgAddValidat
 	return &types.MsgAddValidatorsResponse{}, nil
 }
 
+func (k msgServer) DeleteValidator(goCtx context.Context, msg *types.MsgDeleteValidator) (*types.MsgDeleteValidatorResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	err := k.RemoveValidatorFromHostZone(ctx, msg.HostZone, msg.ValAddr)
+	if err != nil {
+		errMsg := fmt.Sprintf("Validator (%s) not removed from host zone (%s) | err: %s", msg.ValAddr, msg.HostZone, err.Error())
+		k.Logger(ctx).Error(errMsg)
+		return nil, errorsmod.Wrapf(types.ErrValidatorNotRemoved, errMsg)
+	}
+
+	return &types.MsgDeleteValidatorResponse{}, nil
+}
+
 func (k msgServer) ChangeValidatorWeight(goCtx context.Context, msg *types.MsgChangeValidatorWeights) (*types.MsgChangeValidatorWeightsResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
